requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
pandas==2.2.1
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
python-dotenv>=1.0.0
requests>=2.31.0
tqdm==4.66.2 